requirements.txt
# 项目依赖包
# Project dependencies

# 数据处理和验证
pydantic>=1.10.0
pandas>=1.5.0
numpy>=1.21.0
PyYAML>=6.0

# 优化求解
cvxpy>=1.3.0
gurobipy  # 可选，需要商业许可证
pulp>=2.7.0  # 开源替代

# 科学计算
scipy>=1.9.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter支持
jupyter>=1.0.0
ipykernel>=6.0.0

# 测试
pytest>=7.0.0
pytest-cov>=4.0.0